# mlmsittu/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:18080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/items/{id}": {
      "get": {
        "tags": [
          "catalogue-controller"
        ],
        "operationId": "getItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "catalogue-controller"
        ],
        "operationId": "updateItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "catalogue-controller"
        ],
        "operationId": "deleteItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/items": {
      "get": {
        "tags": [
          "catalogue-controller"
        ],
        "operationId": "listItems",
        "parameters": [
          {
            "name": "includeInactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseItemResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalogue-controller"
        ],
        "operationId": "createItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{id}/deactivate": {
      "post": {
        "tags": [
          "catalogue-controller"
        ],
        "operationId": "deactivateItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{id}/activate": {
      "post": {
        "tags": [
          "catalogue-controller"
        ],
        "operationId": "activateItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/item-sets/{id}/deactivate": {
      "post": {
        "tags": [
          "item-set-controller"
        ],
        "operationId": "deactivate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemSetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/item-sets/{id}/activate": {
      "post": {
        "tags": [
          "item-set-controller"
        ],
        "operationId": "activate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemSetResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/categories": {
      "get": {
        "tags": [
          "catalogue-controller"
        ],
        "operationId": "listCategories",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseCategoryResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalogue-controller"
        ],
        "operationId": "createCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateItemRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "unitCost": {
            "type": "number",
            "minimum": 0
          },
          "sellingPrice": {
            "type": "number",
            "minimum": 0
          },
          "retailPrice": {
            "type": "number",
            "minimum": 0
          },
          "wholesalePrice": {
            "type": "number",
            "minimum": 0
          },
          "reorderLevel": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "required": [
          "name",
          "unitCost"
        ]
      },
      "ItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "unitCost": {
            "type": "number"
          },
          "sellingPrice": {
            "type": "number"
          },
          "retailPrice": {
            "type": "number"
          },
          "wholesalePrice": {
            "type": "number"
          },
          "reorderLevel": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ComponentResponse": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ItemSetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "setPrice": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          },
          "imageId": {
            "type": "string",
            "format": "uuid"
          },
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentResponse"
            }
          }
        }
      },
...
      "CreateItemRequest": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "maxLength": 64,
            "minLength": 0
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "unitCost": {
            "type": "number",
            "minimum": 0
          },
          "sellingPrice": {
            "type": "number",
            "minimum": 0
          },
          "retailPrice": {
            "type": "number",
            "minimum": 0
          },
          "wholesalePrice": {
            "type": "number",
            "minimum": 0
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "openingQuantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "reorderLevel": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "required": [
          "name",
          "sku",
          "unitCost"
        ]
      },
...
      "CreateCategoryRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 0
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "PagedResponseItemResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemResponse"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        }
      },
...
      "PagedResponseCategoryResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryResponse"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        }
      },
...
    }
  }
}